Remove disableSSL property from docs
The purpose of this property was to set the "scheme" for the blob
storage endpoint to "http" rather than the implicit "https" when just a
hostname is provided.

This is a rare use case limited to development, so we have removed this
property.  A user using a blob store that does not support TLS should
specify endpoint: "http://my.endpoint/" instead, which we already
demonstrate in the docs.

[#176718452](https://www.pivotaltracker.com/story/show/176718452)