Added option to use older password algo for older clients
When I tried to connect from a TAS rails app to a k8s deployed MySQL
instance, the version of the MariaDB client we ship inside of TAS (from
ubuntu 18.04) is not new enough to support the password hashing
algorithm deployed by this MySQL8 offering. Let's give customers the
option to use older hashing, which I tried and that worked.

Signed-off-by: David Stevenson <stevensonda@vmware.com>